Refactor bridge logic into package modules
