+++ evals/test_live_flexo_export.py
@@ -7,9 +7,9 @@ import uuid
 from pathlib import Path
 
 ROOT = Path(__file__).resolve().parents[1]
-sys.path.insert(0, str(ROOT / "scripts"))
+sys.path.insert(0, str(ROOT / "src"))
 
-import flexo_syson_bridge  # noqa: E402
+from mbse_lab.bridge import workflow as flexo_syson_bridge  # noqa: E402
 
 
 class LiveFlexoExportTests(unittest.TestCase):
