+++ src/mbse_lab/bridge/__init__.py
@@ -1,0 +1,53 @@
+"""Bridge package for Flexo export, SysML rendering, and SysON import."""
+
+from mbse_lab.bridge.workflow import (
+    DEFAULT_FLEXO_URL,
+    DEFAULT_SYSON_URL,
+    DeploymentContract,
+    DeploymentVerificationReport,
+    FlexoSnapshot,
+    build_parser,
+    default_output_dir,
+    deployment_contract_from_snapshot,
+    export_flexo_project,
+    format_deployment_contract_table,
+    format_deployment_verification_report,
+    graphql,
+    load_deployment_env,
+    main,
+    read_json,
+    render_snapshot,
+    request,
+    request_json,
+    run_log_path,
+    sanitize_identifier,
+    verify_deployment_contract,
+    verify_deployment_mounts,
+    write_run_log,
+)
+
+__all__ = (
+    "DEFAULT_FLEXO_URL",
+    "DEFAULT_SYSON_URL",
+    "DeploymentContract",
+    "DeploymentVerificationReport",
+    "FlexoSnapshot",
+    "build_parser",
+    "default_output_dir",
+    "deployment_contract_from_snapshot",
+    "export_flexo_project",
+    "format_deployment_contract_table",
+    "format_deployment_verification_report",
+    "graphql",
+    "load_deployment_env",
+    "main",
+    "read_json",
+    "render_snapshot",
+    "request",
+    "request_json",
+    "run_log_path",
+    "sanitize_identifier",
+    "verify_deployment_contract",
+    "verify_deployment_mounts",
+    "write_run_log",
+)
