+++ src/mbse_lab/deployment/__init__.py
@@ -1,0 +1,25 @@
+"""Deployment runtime verification package."""
+
+from mbse_lab.deployment.verify import (
+    DeploymentContract,
+    DeploymentVerificationReport,
+    deployment_contract_from_snapshot,
+    format_deployment_contract_table,
+    format_deployment_verification_report,
+    load_deployment_env,
+    verify_deployment_contract,
+    verify_deployment_mounts,
+    verify_deployment_ports,
+)
+
+__all__ = (
+    "DeploymentContract",
+    "DeploymentVerificationReport",
+    "deployment_contract_from_snapshot",
+    "format_deployment_contract_table",
+    "format_deployment_verification_report",
+    "load_deployment_env",
+    "verify_deployment_contract",
+    "verify_deployment_mounts",
+    "verify_deployment_ports",
+)
