+++ src/mbse_lab/deployment/verify.py
@@ -1,0 +1,29 @@
+"""Deployment runtime contract verification helpers."""
+
+from mbse_lab.bridge.workflow import (
+    DeploymentContract,
+    DeploymentVerificationReport,
+    deployment_contract_from_snapshot,
+    format_deployment_contract_table,
+    format_deployment_verification_report,
+    inspect_compose_service,
+    inspect_docker_container,
+    load_deployment_env,
+    verify_deployment_contract,
+    verify_deployment_mounts,
+    verify_deployment_ports,
+)
+
+__all__ = (
+    "DeploymentContract",
+    "DeploymentVerificationReport",
+    "deployment_contract_from_snapshot",
+    "format_deployment_contract_table",
+    "format_deployment_verification_report",
+    "inspect_compose_service",
+    "inspect_docker_container",
+    "load_deployment_env",
+    "verify_deployment_contract",
+    "verify_deployment_mounts",
+    "verify_deployment_ports",
+)
